src/main/webapp/app/entities/monthly-revenue/monthly-revenue.component.ts
import { defineComponent, inject, onMounted, ref, type Ref, watch } from 'vue';

import MonthlyRevenueService from './monthly-revenue.service';
import { type IMonthlyRevenue } from '@/shared/model/monthly-revenue.model';
import { useAlertService } from '@/shared/alert/alert.service';

export default defineComponent({
  compatConfig: { MODE: 3 },
  name: 'MonthlyRevenue',
  setup() {
    const monthlyRevenueService = inject('monthlyRevenueService', () => new MonthlyRevenueService());
    const alertService = inject('alertService', () => useAlertService(), true);

    const itemsPerPage = ref(20);
    const queryCount: Ref<number> = ref(null);
    const page: Ref<number> = ref(1);
    const propOrder = ref('id');
    const reverse = ref(false);
    const totalItems = ref(0);

    const monthlyRevenues: Ref<IMonthlyRevenue[]> = ref([]);

    const isFetching = ref(false);

    const clear = () => {
      page.value = 1;
    };

    const sort = (): Array<any> => {
      const result = [propOrder.value + ',' + (reverse.value ? 'desc' : 'asc')];
      if (propOrder.value !== 'id') {
        result.push('id');
      }
      return result;
    };

    const retrieveMonthlyRevenues = async () => {
      isFetching.value = true;
      try {
        const paginationQuery = {
          page: page.value - 1,
          size: itemsPerPage.value,
          sort: sort(),
        };
        const res = await monthlyRevenueService().retrieve(paginationQuery);
        totalItems.value = Number(res.headers['x-total-count']);
        queryCount.value = totalItems.value;
        monthlyRevenues.value = res.data;
      } catch (err) {
        alertService.showHttpError(err.response);
      } finally {
        isFetching.value = false;
      }
    };

    const handleSyncList = () => {
      retrieveMonthlyRevenues();
    };

    onMounted(async () => {
      await retrieveMonthlyRevenues();
    });

    const removeId: Ref<number> = ref(null);
    const removeEntity = ref<any>(null);
    const prepareRemove = (instance: IMonthlyRevenue) => {
      removeId.value = instance.id;
      removeEntity.value.show();
    };
    const closeDialog = () => {
      removeEntity.value.hide();
    };
    const removeMonthlyRevenue = async () => {
      try {
        await monthlyRevenueService().delete(removeId.value);
        const message = 'A MonthlyRevenue is deleted with identifier ' + removeId.value;
        alertService.showInfo(message, { variant: 'danger' });
        removeId.value = null;
        retrieveMonthlyRevenues();
        closeDialog();
      } catch (error) {
        alertService.showHttpError(error.response);
      }
    };

    const changeOrder = (newOrder: string) => {
      if (propOrder.value === newOrder) {
        reverse.value = !reverse.value;
      } else {
        reverse.value = false;
      }
      propOrder.value = newOrder;
    };

    // Whenever order changes, reset the pagination
    watch([propOrder, reverse], async () => {
      if (page.value === 1) {
        // first page, retrieve new data
        await retrieveMonthlyRevenues();
      } else {
        // reset the pagination
        clear();
      }
    });

    // Whenever page changes, switch to the new page.
    watch(page, async () => {
      await retrieveMonthlyRevenues();
    });

    return {
      monthlyRevenues,
      handleSyncList,
      isFetching,
      retrieveMonthlyRevenues,
      clear,
      removeId,
      removeEntity,
      prepareRemove,
      closeDialog,
      removeMonthlyRevenue,
      itemsPerPage,
      queryCount,
      page,
      propOrder,
      reverse,
      totalItems,
      changeOrder,
    };
  },
});
